Pixel loops carry their chunk width in the type
The RGB/RGBA/UTF-16 walks now use `as_chunks::<N>()` and `as_chunks_mut::<N>()` instead of `chunks_exact(N)`, so each chunk is a `&[T; N]` with its width known at compile time. Same pixels, same dropped remainder, but the bounds checks fold away and a wrong index is a compile error rather than a panic.

- 44 call sites across `decoding`, `color`, `histogram`, `clipboard`, `printing`, `previews`, `render`, the Windows browser and settings, plus the RAW examples and tests.
- Two of them needed `.iter_mut()` where clippy's suggestion wrote `.iter()`: the gain-map and warp-rectilinear loops in `dng_opcodes.rs` write through the chunk, and its `enumerate()` rewrite would have handed out shared references. `raw-dev-dump.rs`'s sharpen loop had the same shape.
- `grid.rs` drops a no-op `images.0 as isize`; `HIMAGELIST` already wraps an `isize`.
- `make-app-icon.rs` picks its source with `rfind` rather than `filter().next_back()`.

All of this is what clippy 0.1.98 asks for, and it is what turned CI red on Linux while a Mac on an older `stable` stayed green.

+++ apps/desktop/src/browser/windows/ui/grid.rs
@@ -164,7 +164,7 @@ pub(super) fn create(
             hwnd,
             LVM_SETIMAGELIST,
             Some(WPARAM(LVSIL_NORMAL as usize)),
-            Some(LPARAM(images.0 as isize)),
+            Some(LPARAM(images.0)),
         )
     };
     apply_icon_spacing(hwnd, dpi);
